+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/api/AccountController.java
@@ -1,0 +1,15 @@
+package com.windea.demo.cloudcollect.core.api;
+
+import com.windea.demo.cloudcollect.core.service.UserService;
+import org.springframework.web.bind.annotation.*;
+
+@CrossOrigin
+@RestController
+@RequestMapping("/account")
+public class AccountController {
+	private final UserService userService;
+
+	public AccountController(UserService userService) {
+		this.userService = userService;
+	}
+}