+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/api/CollectController.java
@@ -1,0 +1,15 @@
+package com.windea.demo.cloudcollect.core.api;
+
+import com.windea.demo.cloudcollect.core.service.CollectService;
+import org.springframework.web.bind.annotation.*;
+
+@CrossOrigin
+@RestController
+@RequestMapping("/collect")
+public class CollectController {
+	private final CollectService service;
+
+	public CollectController(CollectService service) {
+		this.service = service;
+	}
+}