+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/domain/entity/Collect.java
@@ -1,0 +1,82 @@
+package com.windea.demo.cloudcollect.core.domain.entity;
+
+import com.windea.demo.cloudcollect.core.domain.enums.CollectPrivacy;
+import com.windea.demo.cloudcollect.core.domain.enums.CollectType;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.lang.Nullable;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.Size;
+import java.io.Serializable;
+import java.time.LocalDateTime;
+import java.util.*;
+
+/**
+ * 收藏。
+ */
+@Data
+@NoArgsConstructor
+@Entity
+public class Collect implements Serializable {
+	private static final long serialVersionUID = -6764369348818887548L;
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@NotEmpty(message = "validation.Collect.title.NotEmpty")
+	@Size(min = 1, max = 64, message = "validation.Collect.title.Size")
+	@Column(nullable = false, length = 64)
+	private String title;
+
+	@Column(nullable = false, length = 512)
+	private String url;
+
+	@Nullable
+	@Column(length = 512)
+	private String logoUrl;
+
+	@NotEmpty(message = "validation.Collect.summary.NotEmpty")
+	@Size(min = 1, max = 255, message = "validation.Collect.summary.Size")
+	@Column(nullable = false, columnDefinition = "text")
+	private String summary;
+
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false)
+	private CollectCategory category;
+
+	@ManyToMany(cascade = CascadeType.MERGE, fetch = FetchType.EAGER)
+	@JoinColumn
+	private Set<CollectTag> tags = new LinkedHashSet<>();
+
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false)
+	private CollectType type;
+
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false)
+	private CollectPrivacy privacy;
+
+	@OneToMany(cascade = CascadeType.MERGE, fetch = FetchType.EAGER)
+	@JoinColumn
+	private List<Comment> commentList = new LinkedList<>();
+
+	@ManyToOne(cascade = CascadeType.MERGE, fetch = FetchType.EAGER)
+	@JoinColumn(nullable = false)
+	private User user;
+
+	@CreatedDate
+	@Column
+	private LocalDateTime createdTime;
+
+	@LastModifiedDate
+	@Column
+	private LocalDateTime lastModifiedTime;
+
+	@Column(nullable = false)
+	private Boolean deleted = false;
+}