+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/domain/entity/Praise.java
@@ -1,0 +1,42 @@
+package com.windea.demo.cloudcollect.core.domain.entity;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.time.LocalDateTime;
+import java.util.LinkedList;
+import java.util.List;
+
+/**
+ * 点赞信息。
+ */
+@Data
+@NoArgsConstructor
+@Entity
+public class Praise implements Serializable {
+	private static final long serialVersionUID = -3754697767735093276L;
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@OneToOne(cascade = CascadeType.MERGE, fetch = FetchType.EAGER)
+	@JoinColumn(nullable = false)
+	private Collect collect;
+
+	@OneToMany(cascade = CascadeType.MERGE, fetch = FetchType.EAGER)
+	@JoinColumn
+	private List<User> praisedUserList = new LinkedList<>();
+
+	@CreatedDate
+	@Column
+	private LocalDateTime createdTime;
+
+	@LastModifiedDate
+	@Column
+	private LocalDateTime lastModifiedTime;
+}