+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/domain/enums/CollectType.java
@@ -1,0 +1,12 @@
+package com.windea.demo.cloudcollect.core.domain.enums;
+
+/**
+ * 收藏的类型。
+ */
+public enum CollectType {
+	NONE,
+	LOVE,
+	IMPORTANT,
+	TODO,
+	DELAY
+}