+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/domain/enums/NoticeType.java
@@ -1,0 +1,10 @@
+package com.windea.demo.cloudcollect.core.domain.enums;
+
+/**
+ * 通知的类型。
+ */
+public enum NoticeType {
+	NEW_COMMENT,
+	NEW_FOLLOW,
+	NEW_PRAISE
+}