+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/domain/view/UserLoginView.java
@@ -1,0 +1,9 @@
+package com.windea.demo.cloudcollect.core.domain.view;
+
+import java.io.Serializable;
+
+/**
+ * 用户登录视图接口。
+ */
+public interface UserLoginView extends Serializable {
+}