+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/repository/CollectTagRepository.java
@@ -1,0 +1,7 @@
+package com.windea.demo.cloudcollect.core.repository;
+
+import com.windea.demo.cloudcollect.core.domain.entity.CollectTag;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CollectTagRepository extends JpaRepository<CollectTag, Long> {
+}