+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/repository/PraiseRepository.java
@@ -1,0 +1,7 @@
+package com.windea.demo.cloudcollect.core.repository;
+
+import com.windea.demo.cloudcollect.core.domain.entity.Praise;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PraiseRepository extends JpaRepository<Praise, Long> {
+}