+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/service/FollowService.java
@@ -1,0 +1,4 @@
+package com.windea.demo.cloudcollect.core.service;
+
+public interface FollowService {
+}