+++ cloud-collect-core/src/main/java/com/windea/demo/cloudcollect/core/service/impl/EmailServiceImpl.java
@@ -1,0 +1,8 @@
+package com.windea.demo.cloudcollect.core.service.impl;
+
+import com.windea.demo.cloudcollect.core.service.EmailService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class EmailServiceImpl implements EmailService {
+}